Fix my_inject and delete prueba.rb
